+++ enemy.py
@@ -17,8 +17,8 @@ class Enemy:
     def is_alive(self):
         return self.health > 0
 
-    def can_cast(self):
-        return self.mana > 0
+    def can_cast(self, spell):
+        return self.mana >= spell.mana_cost
 
     def get_health(self):
         return self.health
@@ -26,24 +26,17 @@ class Enemy:
     def get_mana(self):
         return self.mana
 
-    def take_damage(self, damage):
-        self.health -= damage
+    def take_damage(self, damage_points):
+        self.health = max(0, self.health - damage_points)
 
     def take_healing(self, healing_points):
-        if not self.is_alive():
-            return False
-        else:
-            if self.health + healing_points > 100:
-                self.health = 100
-            else:
-                self.health += healing_points
+        if self.is_alive():
+            self.health = min(100, self.health + healing_points)
             return True
+        return False
 
     def take_mana(self, mana_points):
-        if self.mana + mana_points > 100:
-            self.mana = 100
-        else:
-            self.mana += mana_points
+        self.mana = min(self.mana_save, self.mana + mana_points)
 
     def attack(self):
         return self.damage
