+++ fight.py
@@ -1,0 +1,4 @@
+class Fight:
+    def __init__(self, hero, enemy):
+        self.hero = hero
+        self.enemy = enemy
