+++ hero.py
@@ -2,73 +2,81 @@ from weapon import Weapon, Spell
 
 
 class Hero:
-    def __init__(self, name="Bron",\
-                 title="Dragonslayer", health=100,\
+    def __init__(self, name="Bron", title="Dragonslayer", health=100,
                  mana=100, mana_regeneration_rate=2):
         assert type(name) is str
         assert type(title) is str
         assert type(health) is float or type(health) is int
         assert type(mana) is float or type(mana) is int
         assert type(mana_regeneration_rate) is float or\
-               type(mana_regeneration_rate) is int
+            type(mana_regeneration_rate) is int
         if health < 0 or mana < 0 or mana_regeneration_rate < 0:
             raise ValueError
         if health > 100 or mana > 100:
             raise ValueError
         self.name = name
         self.title = title
-        self._health = health
-        self._mana = mana
+        self.health = health
+        self.mana = mana
         self.mana_regeneration_rate = mana_regeneration_rate
-        self._weapon = None
-        self._spell = None
+        self.mana_save = mana
+        self.weapon = None
+        self.spell = None
+        self.weapon_inventory = [Weapon('The Axe Of Destiny', 20)]
+        self.spell_inventory = [Spell(name='Fireball',
+                                      damage=30,
+                                      mana_cost=50,
+                                      cast_range=2)]
 
     def known_as(self):
         return f"{self.name} the {self.title}"
 
     def is_alive(self):
-        return self._health > 0
-
-    def can_cast(self):
-        return self._mana > 0
+        return self.health > 0
 
     def get_health(self):
-        return self._health
+        return self.health
 
     def get_mana(self):
-        return self._mana
+        return self.mana
 
-    def take_healing(self, healing_points):
-        if not self.is_alive():
-            return False
-        if self.get_health() + healing_points > 100:
-            return False
-        self._health += healing_points
-        return True
-
-    def take_mana(self, mana_points):
-        if self.get_mana() + self.mana_regeneration_rate + mana_points < 100:
-            self._mana = self.get_mana() + self.mana_regeneration_rate + mana_points
-
-    def attack(self, by):
-        if self._weapon is None and self._spell is None:
-            return 0
-        if by == "weapon":
-            return self._weapon.get_damage()
-        if by == "magic":
-            return self._spell.get_damage()
-
-    def equip(self, weapon):
-        assert type(weapon) is Weapon
-        self._weapon = weapon
-
-    def learn(self, spell):
-        assert type(spell) is Spell
-        self._spell = spell
+    def can_cast(self, spell):
+        return spell.mana_cost <= self.mana
 
     def take_damage(self, damage_points):
-        if self.get_health() - damage_points < 0:
-            self._health = 0
-        else:
-            self._health -= damage_points
+        self.health = max(0, self.health - damage_points)
 
+    def take_healing(self, healing_points):
+        if self.is_alive():
+            self.health = min(100, self.health + healing_points)
+            return True
+        return False
+
+    def take_mana(self, mana_points):
+        self.mana = min(self.mana_save, self.mana + mana_points)
+
+    def equip(self, weapon):
+        if weapon in self.weapon_inventory:
+            self.weapon = weapon
+        else:
+            print('You do not have that weapon in your weapon inventory!')
+
+    def learn(self, spell):
+        if spell in self.spell_inventory:
+            self.spell = spell
+        else:
+            print("You do not have that spell in your spell inventory! ")
+
+    def attack(self, by):
+        if by == 'weapon':
+            return self.weapon.damage if self.weapon is not None else 0
+        if by == 'spell':
+            return self.spell.damage if self.damage is not None else 0
+
+    def get_weapon_inventory(self):
+        for weapon in self.weapon_inventory:
+            print(weapon)
+
+    def get_spell_inventory(self):
+        for spell in self.spell_inventory:
+            print(spell)
